# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Analizcim API",
    "description": "Excel analiz uygulaması için REST API. Bu API, satış ve alış dosyalarını analiz etmek, geçmişi yönetmek, tahminler oluşturmak ve kullanıcı yönetimi için kullanılır.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@analizcim.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Yerel geliştirme sunucusu"
    }
  ],
...
  "paths": {
...
    "/expenses-local": {
      "get": {
        "tags": ["Giderler"],
        "summary": "Giderleri listele",
        "description": "Sabit ve değişken giderleri listeler.",
        "operationId": "getExpenses",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filtreleme için yıl"
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            },
            "description": "Filtreleme için ay"
          }
        ],
        "responses": {
          "200": {
            "description": "Gider verileri",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "year": {
                          "type": "string"
                        },
                        "month": {
                          "type": "string"
                        },
                        "fixed": {
                          "type": "array"
                        },
                        "variable": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Giderler"],
        "summary": "Giderleri güncelle",
        "description": "Sabit ve değişken giderleri kaydeder/günceller.",
        "operationId": "updateExpenses",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["year"],
                "properties": {
                  "year": {
                    "type": "integer"
                  },
                  "month": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  },
                  "fixed": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "category": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "variable": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "category": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Giderler kaydedildi"
          }
        }
      }
    },
    "/expenses-local/migrate": {
      "post": {
        "tags": ["Giderler"],
        "summary": "Giderleri taşı",
        "description": "Giderleri farklı bir kaynaktan içe aktarır.",
        "operationId": "migrateExpenses",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "category": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string",
                          "enum": ["fixed", "variable"]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Giderler taşındı"
          }
        }
      }
    },
    "/expenses-local/years": {
      "get": {
        "tags": ["Giderler"],
        "summary": "Veri olan yılları getir",
        "operationId": "getExpenseYears",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Yıllar listesi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "years": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Oturum kimlik doğrulaması"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Admin yetkisi gerektiren işlemler için"
      }
    }
  }
}